Document expectations for commit messages
This commit tries to be more explicit about the expectations for commit
messages and testing within bmcweb, and give maintainers something to
point to instead of repeating the same statements over and over again.

This will likely need to evolve, but I think it's a good start, and
would help people put together commits that can be merged on the first
try, rather than requiring followup.

Tested: Documentation only, no testing.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Ib60ac90eccdc4f2c3215a755bbb462fff4796dd4