Make OWNERS match MAINTAINERS
Krzysztof was added as a reviewer in the MAINTAINERS file but not OWNERS
file as part of 4576ec1.

Add Krzysztof as a reviewer in OWNERS.

One day this should mean Krzysztof is automatically added to bmcweb
reviews, after support for automatically adding Reviewers is added.

Copied syntax from other repos OWNERS. E.g.
https://github.com/openbmc/docs/blob/master/OWNERS#L39

Change-Id: I6b721fe5d6eb1c45a9476c2b89d25b6998f6bb1d
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>