host-state: do not return anything if unavailable
The host state information is provided by the
xyz.openbmc_project.State.Host service. There is no guarantee that this
service will be up and running by the time bmcweb needs it. Returning an
InternalError simply because a service is not yet running is
not very user friendly to our clients. In most situations, a client will
ignore all data returned when a 500 is returned.

Instead of putting systemd Wants/Before type relationships on everything
possibly needed by bmcweb, the design point is to simply return what
bmcweb can get at that instant in time. With this change, the majority
of the redfish system object data can be returned and used by the
client.

This scenario has been seen a few times on our p10bmc machine.

Tested:
- Verified that when xyz.openbmc_project.State.Host was unavailable, a
  call to redfish/v1/Systems/system returned what was available without
  error.
- Verified that redfish validator passed
- Verified that redfish validator passed when
  xyz.openbmc_project.State.Host.service was unavailable

Change-Id: I22c6942d2c81083bf90fa4180e95b1fa19221374
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>