Make parse_registries do whitespace properly
Previously, parse_registries was not very careful about generating
readable files, and relied on clang-format to "fix" them after the fact.
Given they're generated, this is unforunate, and leads to some
inconsistencies in reading the generated code.

This commit changes the script to no longer rely on clang-format, and
wrap the whole file in a clang-format off bracket.  This means that our
message registry generation will be consistent.

As an added bonus, different versions of clang-format like to format
these structures differently, so this removes one possible barrier to
people updating these.

This was initiated by the next patch in the series, which made a
seemingly minor change to a comment, which caused a significant change
in this file because of clang-format versioning.

Tested: ran parse_registries.py and saw build passing.  Whitespace
changes only.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Id48bb43dd9d8ecc83de1840f2f1045a87e2ff796