Don't rely on operator << for object logging
In the upcoming fmt patch, we remove the use of streams, and a number of
our logging statements are relying on them.  This commit changes them to
no longer rely on operator>> or operator+ to build their strings.  This
alone isn't very useful, but in the context of the next patch makes the
automation able to do a complete conversion of all log statements
automatically.

Tested: enabled logging on local and saw log statements print to console

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I0e5dc2cf015c6924037e38d547535eda8175a6a1