auth: change authorization.hpp to authentication.hpp
The existing authorization header is actually doing "authentication"
work. The authorization is happening in routing.hpp where we fetch the
role of the authenticated user and get their privilege set.

This commits changes the name of the file, as well as the namespace, to
be more precise on what the file actually does.

Tested:
1. Trivial change, it builds

Signed-off-by: Nan Zhou <nanzhoumails@gmail.com>
Change-Id: Ib91ed70507a7308522c7e5363ed2f4dc279a19d9