kvm_websocket_test: delete the test
This test is not enabled and referenced anywhere. Per Discord
discussion, this test is obsolete and can be deleted now.
https://discord.com/channels/775381525260664832/855566794994221117/985996960840429568

"""
bmcweb used to have the full RFB (VNC) server in it, and access the
linux device directly, that's what that unit test was checking, but
that code is long gone in lieu of what we have now (unix socket to the
rfbserver) which is better.
"""

Signed-off-by: Nan Zhou <nanzhoumails@gmail.com>
Change-Id: If42801c15009f8b33ea5d15749a067dccda935e9