Include Collections for JsonSchema resources
Redfish collection Json schema files are not versioned, and only have a
single version file.  In the script, we assumed that all files were
versioned, and the script ignored files that didn't parse as versioned.

This commit fixes the script, and checks in the relevant Collection
resources.

Tested: Redfish service validator passes on JsonSchemas resources and
collection.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I579dac3fdfecf485f7d414842effa9ae0ee517e8