HTTP Client: Fix handling on connection timeout
If a destination is not reachable, then the connection will timeout
in doConnect().  This causes issues later when the client attempts to
resend the message.

The error check in doClose() should not exit early since that will
result in the connection's status not being marked as closed and
thus it will never get reused.

Similarly, doCloseAndRetry() should not exit early since that will
cause the retry flow to hang and the connection's callback function
will not get deleted.

Tested:
Used Redfish Aggregation patches in the chain through
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/54896 to verify that
requests to collections such as /redfish/v1/Chassis no longer hang
when the specified Satellite BMC does not exist

Signed-off-by: Carson Labrado <clabrado@google.com>
Change-Id: Ic8369b0de8efb00ff168bc1ed43f1d7fd6c7366a