IBM Mgmt console writeLock conflict fix
The isConflictRecord method was returning false by default which is
causing ignoring the same resourceId conflicts.
Same resourceId case will pass all the conditions and reach to the end
of the function. Returning true means that there is a conflict.

This commit fixes this by returning true by default

Tested by:
1. Send writeLock requests with same resourceId and segment length

Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>
Change-Id: Ie65c6394988a357a8c811b621e113c14924bb8f6