Allow parsing urls with extra elements
Sometimes it is desirable to only parse a portion of a URL, and allow
any elements at the end.  This comes up significantly in aggregation
where parsing "/redfish/v1/<collection>/anything is pretty common.

This commit adds a new class, OrMorePaths, that can be used as a
placeholder to indicate that more paths should be accepted.

Tested: Unit tests pass.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: If4fb3991a91560fd3b8b838f912aa36e79ddd2b3