Run black formatter on scripts/update_schemas.py
https://black.readthedocs.io/en/stable/

Is a formatter that takes a more opinionated stance on pep8 compliance
than others, and as such gives us more consistent python code, even if
that's not currently enforced by ci.  In addition, it allows patches to
automate their pep8 compliance similar to how we've done clang-format
for c++.  We've done this transform other places throughout the codebase
and gotten a decent result, so duplicate it in bmcweb..

Run the black tool on update_schemas.py.

Tested: Reran update_schemas.py, and generated output is identical

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I83f47025ac61d0afd5df87ddb33f84269160191c