Fix json-array copy and refactor assembly
A json array is mistakenly copied as a part of [1]. This is to remove
the unnecessary copy operation but use const reference.

```
       nlohmann::json::array_t assemblyArray =
-            asyncResp->res.jsonValue["Assemblies"];

+        nlohmann::json& assemblyArray = asyncResp->res.jsonValue["Assemblies"];
```

This commit also refactors and simplifies a few aux functions
to get the chassis assembly into an util function.
- getChassisAssembly()

It is done by uding `getAssociatedSubTreePaths()`.
This util function can also be used in other places.

[1] https://github.com/ibm-openbmc/bmcweb/blame/3ca5028e701457d1a0a454bbde2ec75b62ee87c9/redfish-core/lib/assembly.hpp#L194-L197C34

Signed-off-by: Shantappa Teekappanavar <shantappa.teekappanavar@ibm.com>
Signed-off-by: Myung Bae <myungbae@us.ibm.com>