Make Request copy explicit
It is currently too easy to accidentally make copies of the Request
object.  Ideally code would parse out the Request in the first handler,
then no longer require an async copy.  There is one case in the redfish
query things where we actually need a copy of the request object, so we
need these constructors, but we should make them explicit.

This commit moves the Request constructor to be private, and adds a new
method called copy() for explicitly making a copy.  Ironcially, this
finds one place where we were actually making a copy of the request
object unintentionally, so fix that to only capture the value
required,the user session.

Tested:
- Compiles
- Run GET/PATCH related curl or If-Match like PATCH Account
- Redfish Service Validator runs and passes

Change-Id: I19255981f42757ed736112c003201e3f758735ac
Signed-off-by: Ed Tanous <ed@tanous.net>

2. Fix build break

608ad2cc0e17bf62ff34bae7ceb499816fe13fbb
and
c1a75ebc267a78853fb26a3da8c6b3388e6ee07c
Collided.  Fix the build error.

Change-Id: I683ea7a7b21b2297ab93010815b01e04373e1cea
Signed-off-by: Ed Tanous <ed@tanous.net>

3. Apply the changes for additional files for downstream

Signed-off-by: Myung Bae <myungbae@us.ibm.com>