Use SPDX identifiers everywhere
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/74513 moved most places
over to SPDX but forgot a few, fix the ones found with grep.

SPDX identifiers are simpler, and reduce the amount of cruft we have in
code files.  They are recommended by linux foundation, and therefore we
should do as they allow.

This patchset does not intend to modify any intent on any existing
copyrights or licenses, only to standardize their inclusion.

[1] https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects

Change-Id: Ie4c2ea53f7bc8d902bf87fef6df2a67c6b1de613
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>