Managers: Factor out large lambdas
Restructure to move large lambdas to named functions for getting the
properties for the manager. No semantic changes were made.

Note: I refactored only some of the large lambdas in this file and path.
The ones I refactored are in the flow for a follow-on commit to add the
LocationIndicatorActive Redfish property to the Manager.

```
curl -k -H "X-Auth-Token: $token" -X GET https://${bmc}/redfish/v1/Managers/bmc
{
  "@odata.id": "/redfish/v1/Managers/bmc",
  "@odata.type": "#Manager.v1_14_0.Manager",
  ...
}
```

Tested: Confirmed same output before and after change.

Change-Id: I350ecbc5add46cfc6356d05b4feb484bb448a323
Signed-off-by: Janet Adkins <janeta@us.ibm.com>