1120: Ship and install all schemas (#910) (#1329)
* Fix json-schema-installed version during schema update

This fixes `update_schemas.py` so that the existing installed csdl and
json symlinks are to be kept and also to be updated with the matching
json schema version when DMTF schema version is upgraded.

This commit also uses the symlinks to the closed relative paths like
- `../schema/dmtf/installed/<schema>.xml -> ../csdl/<schema>.xml`
- `../schema/dmtf/json-schema-installed/<json-file>.json ->
  ../json-schema/<json-file>.json`

Tested:
 - Change VERSION in `update_schemas.py` and verify the generated files
 - CI passes
 - Redfish Service Validator passes

Change-Id: Ib7fede7e4c39bdfc13da227eb1e737d96b6c2b5e
Signed-off-by: Myung Bae <myungbae@us.ibm.com>

* Ship and install all schemas (#910)

This commit is to include and ship all available schemas, as how it was
done in the previous releases.

After modification of `update_schemas.py`, the schema files are
re-generated and they will be included.

Tested:
- CI passes
- GET /redfish/v1/JsonSchemas to see all json schemas
- Redfish Service Validator passes

Signed-off-by: Myung Bae <myungbae@us.ibm.com>

---------

Signed-off-by: Myung Bae <myungbae@us.ibm.com>