dump: Refine socket cleanup & async handling (#1381)
Currently, in the dump_offload code, 'adaptor.lowest_layer()' API is
used while closing the underlying tcp socket. This works for ssl stream,
but in websocket stream, the API will not be able to get the underlying
socket, leading to a stale open socket. This way, for all the dumps
offloaded, the same number of sockets were left stale.

This commit fixes the issue by utilizing "get_lowest_layer()" which is a
generic API that works for ssl stream as well as websocket stream
irrespective of how many wrappers are there around the tcp socket.

It also contains the following changes:
* Updated lambda captures to include 'self(shared_from_this())' to
  ensure proper lifetime management of the handler object.
* Proper closing of unix sockets to prevent resource leaks.

Tested By:
Tested dump offload and it was successful, and no stale unix sockets
left open.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>